+++ src/Compute/Compute.Test/ScenarioTests/VirtualMachineScaleSetTests.cs
@@ -191,6 +191,12 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
         public void TestVirtualMachineScaleSetOrchestrationModeFaultDomain()
         {
             TestRunner.RunTestScript("Test-VirtualMachineScaleSetOrchestrationModeFaultDomain");
+      
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestVirtualMachineScaleSetAssignedHost()
+        {
+            TestRunner.RunTestScript("Test-VirtualMachineScaleSetAssignedHost");
         }
     }
 }