+++ src/Compute/Compute/Strategies/ComputeRp/VirtualMachineStrategy.cs
@@ -54,10 +54,11 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
             Func<IEngine, SubResource> proximityPlacementGroup,
             string hostId,
             string VmssId,
+            string hostGroupId,
             string priority,
             string evictionPolicy,
             double? maxPrice,
-            bool? encryptionAtHostEnabled)
+            bool encryptionAtHostPresent)
 
             => Strategy.CreateResourceConfig(
                 resourceGroup: resourceGroup,
@@ -96,10 +97,11 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
                     ProximityPlacementGroup = proximityPlacementGroup(engine),
                     Host = string.IsNullOrEmpty(hostId) ? null : new SubResource(hostId),
                     VirtualMachineScaleSet = string.IsNullOrEmpty(VmssId) ? null : new SubResource(VmssId),
+                    HostGroup = string.IsNullOrEmpty(hostGroupId) ? null : new SubResource(hostGroupId),
                     Priority = priority,
                     EvictionPolicy = evictionPolicy,
                     BillingProfile = (maxPrice == null) ? null : new BillingProfile(maxPrice),
-                    SecurityProfile = (encryptionAtHostEnabled == null) ? null : new SecurityProfile(encryptionAtHostEnabled)
+                    SecurityProfile = (encryptionAtHostPresent == true) ? new SecurityProfile(encryptionAtHostPresent) : null
                 });
 
         public static ResourceConfig<VirtualMachine> CreateVirtualMachineConfig(
@@ -117,10 +119,11 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
             Func<IEngine, SubResource> proximityPlacementGroup,
             string hostId,
             string VmssId,
+            string hostGroupId,
             string priority,
             string evictionPolicy,
             double? maxPrice,
-            bool? encryptionAtHostEnabled
+            bool encryptionAtHostPresent
             )
             => Strategy.CreateResourceConfig(
                 resourceGroup: resourceGroup,
@@ -156,13 +159,11 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
                     ProximityPlacementGroup = proximityPlacementGroup(engine),
                     Host = string.IsNullOrEmpty(hostId) ? null : new SubResource(hostId),
                     VirtualMachineScaleSet = string.IsNullOrEmpty(VmssId) ? null : new SubResource(VmssId),
+                    HostGroup = string.IsNullOrEmpty(hostGroupId) ? null : new SubResource(hostGroupId),
                     Priority = priority,
                     EvictionPolicy = evictionPolicy,
                     BillingProfile = (maxPrice == null) ? null : new BillingProfile(maxPrice),
-                    SecurityProfile = new SecurityProfile
-                    {
-                        EncryptionAtHost = encryptionAtHostEnabled
-                    }
+                    SecurityProfile = (encryptionAtHostPresent == true) ? new SecurityProfile(encryptionAtHostPresent) : null
                 });
     }
 }