+++ src/Compute/Compute/VirtualMachine/Operation/NewAzureVMCommand.cs
@@ -262,7 +262,15 @@ namespace Microsoft.Azure.Commands.Compute
             HelpMessage = "EncryptionAtHost property can be used by user in the request to enable or disable the Host Encryption for the virtual machine. This will enable the encryption for all the disks including Resource/Temp disk at host itself.")]
         [Parameter(ParameterSetName = DiskFileParameterSet, Mandatory = false,
             HelpMessage = "EncryptionAtHost property can be used by user in the request to enable or disable the Host Encryption for the virtual machine. This will enable the encryption for all the disks including Resource/Temp disk at host itself.")]
-        public SwitchParameter EncryptionAtHost { get; set; } = false;
+        public SwitchParameter EncryptionAtHost { get; set; }
+        
+        [Parameter(ParameterSetName = SimpleParameterSet, Mandatory = false,
+            HelpMessage = "The resource id of the dedicated host group, on which the customer wants their VM placed using automatic placement.",
+            ValueFromPipelineByPropertyName = true)]
+        [Parameter(ParameterSetName = DiskFileParameterSet, Mandatory = false,
+            HelpMessage = "The resource id of the dedicated host group, on which the customer wants their VM placed using automatic placement.",
+            ValueFromPipelineByPropertyName = true)]
+        public string HostGroupId { get; set; }
 
         public override void ExecuteCmdlet()
         {
@@ -397,10 +405,11 @@ namespace Microsoft.Azure.Commands.Compute
                         proximityPlacementGroup: ppgSubResourceFunc,
                         hostId: _cmdlet.HostId,
                         VmssId: _cmdlet.VmssId,
+                        hostGroupId: _cmdlet.HostGroupId,
                         priority: _cmdlet.Priority,
                         evictionPolicy: _cmdlet.EvictionPolicy,
                         maxPrice: _cmdlet.IsParameterBound(c => c.MaxPrice) ? _cmdlet.MaxPrice : (double?)null,
-                        encryptionAtHostEnabled: _cmdlet.EncryptionAtHost.IsPresent
+                        encryptionAtHostPresent: _cmdlet.EncryptionAtHost.IsPresent
                         );
                 }
                 else
@@ -423,11 +432,12 @@ namespace Microsoft.Azure.Commands.Compute
                         proximityPlacementGroup: ppgSubResourceFunc,
                         hostId: _cmdlet.HostId,
                         VmssId: _cmdlet.VmssId,
+                        hostGroupId: _cmdlet.HostGroupId, 
                         priority: _cmdlet.Priority,
                         evictionPolicy: _cmdlet.EvictionPolicy,
                         maxPrice: _cmdlet.IsParameterBound(c => c.MaxPrice) ? _cmdlet.MaxPrice : (double?)null,
-                        encryptionAtHostEnabled: _cmdlet.EncryptionAtHost.IsPresent
-                        );
+                        encryptionAtHostPresent: _cmdlet.EncryptionAtHost.IsPresent
+                    );
                 }
             }
         }