+++ src/Resources/Resources/ChangeLog.md
@@ -19,6 +19,8 @@
 -->
 ## Upcoming Release
 * Added breaking change attribute to `SubscriptionId` parameter of `Get-AzResourceGroupDeploymentOperation`.
+* Updated ARM template What-If cmdlets to show "Ignore" resource changes last.
+* Fixed an issue where dynamic secure string parameters for deployment cmdlets are not correctly serialized.
 
 ## Version 2.5.0
 * Updated `Get-AzPolicyAlias` response to include information indicating whether the alias is modifiable by Azure Policy.